add slider with 5 image and taglines

+++ src/Pages/MyListing.jsx
@@ -2,6 +2,7 @@ import React, { useContext, useEffect, useState } from 'react';
 import { AuthContext } from '../Provider/AuthProvider';
 import { Link } from 'react-router';
 import axios from 'axios';
+import Swal from 'sweetalert2';
 
 const MyListing = () => {
     const [mylisting, setMyListing] = useState([]);
@@ -24,18 +25,40 @@ const MyListing = () => {
     console.log(mylisting);
 
     const hanndleDelete = (id) => {
-        axios.delete(`http://localhost:3000/delete/${id}`)
-            .then(res => {
-                console.log(res.data);
-                const filterData = mylisting.filter(listing => listing._id != id)
-                console.log(filterData);
-                setMyListing(filterData)
 
-            })
-            .catch(err => {
-                console.log(err);
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#3085d6",
+            cancelButtonColor: "#d33",
+            confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                axios.delete(`http://localhost:3000/delete/${id}`)
+                    .then(res => {
+                        console.log(res.data);
+                        if (res.data.deletedCount == 1) {
+                            const filterData = mylisting.filter(listing => listing._id != id)
+                            console.log(filterData);
+                            setMyListing(filterData)
+                            Swal.fire({
+                                title: "Deleted!",
+                                text: "Your file has been deleted.",
+                                icon: "success"
+                            });
+                        }
 
-            })
+                    })
+                    .catch(err => {
+                        console.log(err);
+
+                    })
+
+            }
+        });
+
     }
 
     return (
